+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/gitlab/GitlabLintResponse.kt
@@ -2,6 +2,7 @@ package com.github.blarc.gitlab.template.lint.plugin.gitlab
 
 import kotlinx.serialization.Serializable
 
+@Suppress("PROVIDED_RUNTIME_TOO_LOW")
 @Serializable
 data class GitlabLintResponse(
     val valid: Boolean,