+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/gitlab/GitlabProject.kt
@@ -3,6 +3,7 @@ package com.github.blarc.gitlab.template.lint.plugin.gitlab
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+@Suppress("PROVIDED_RUNTIME_TOO_LOW")
 @Serializable
 data class GitlabProject(
     val id: Long,