+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/Pass.kt
@@ -10,8 +10,10 @@ class Pass(val project: Project, val file: PsiFile) {
     var remote: GitRemote? = null
     var remoteId: Long? = null
     var gitlabUrl: String? = null
+    var gitlabToken: String? = null
 
     fun gitlabUrlOrThrow() = gitlabUrl ?: throw IllegalStateException("Gitlab url not set")
+    fun gitlabTokenOrThrow() = gitlabToken ?: throw IllegalStateException("Gitlab token not set")
     fun repositoryOrThrow() = repository ?: throw IllegalStateException("Repository not set")
     fun remoteOrThrow() = remote ?: throw IllegalStateException("Remote not set")
     fun remoteIdOrThrow() = remoteId ?: throw IllegalStateException("Remote ID not set")