+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/middleware/LintContext.kt
@@ -5,6 +5,7 @@ import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabLintResponse
 import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pass
 import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusEnum
 import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
 
 @Service
 class LintContext : Middleware {
@@ -14,11 +15,24 @@ class LintContext : Middleware {
 
     override fun invoke(pass: Pass, next: () -> Pair<GitlabLintResponse?, LintStatusEnum>?): Pair<GitlabLintResponse?, LintStatusEnum>? {
         val remoteId = pass.remoteIdOrThrow()
-        val branchName = pass.repositoryOrThrow().currentBranchName ?: return null
+        val branch = pass.repositoryOrThrow().currentBranchName ?: return null
 
-        val gitlab = Gitlab(pass.project)
+        val gitlabUrl = pass.gitlabUrlOrThrow()
+        val gitlabToken = pass.gitlabTokenOrThrow()
 
-        gitlabLintResponse = gitlab.lintContent(pass.file.text, remoteId, branchName, pass.project, showGitlabTokenNotification).get()
+        val gitlab = pass.project.service<Gitlab>()
+
+
+        gitlabLintResponse = gitlab.lintContent(
+            gitlabUrl,
+            gitlabToken,
+            pass.file.text,
+            remoteId,
+            branch,
+            showGitlabTokenNotification
+        ).get()
+
+
         val lintStatus = if (gitlabLintResponse?.valid == true) {
             showGitlabTokenNotification = true
             LintStatusEnum.VALID