+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/middleware/ResolveGitlabToken.kt
@@ -1,5 +1,6 @@
 package com.github.blarc.gitlab.template.lint.plugin.pipeline.middleware
 
+import com.github.blarc.gitlab.template.lint.plugin.gitlab.Gitlab
 import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabLintResponse
 import com.github.blarc.gitlab.template.lint.plugin.notifications.Notification
 import com.github.blarc.gitlab.template.lint.plugin.notifications.sendNotification
@@ -7,6 +8,7 @@ import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pass
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusEnum
 import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
 
 @Service
 class ResolveGitlabToken : Middleware {
@@ -14,11 +16,15 @@ class ResolveGitlabToken : Middleware {
     private var showGitlabTokenNotification = true
 
     override fun invoke(pass: Pass, next: () -> Pair<GitlabLintResponse?, LintStatusEnum>?): Pair<GitlabLintResponse?, LintStatusEnum>? {
-        getGitlabToken(pass) ?: return null
+
+        val gitlabToken = resolveGitlabToken(pass) ?: return null
+
+        pass.gitlabToken = gitlabToken
+
         return next()
     }
 
-    private fun getGitlabToken(pass: Pass) : String? {
+    private fun resolveGitlabToken(pass: Pass) : String? {
         val gitlabUrl = pass.gitlabUrlOrThrow()
         val gitlabToken = AppSettings.instance?.getGitlabToken(gitlabUrl)